feature: add configuration to sqllint
