+++ pkg/sqllint/testdata/config.yaml
@@ -1,0 +1,95 @@
+allowed_ddl:
+    create_database_stmt: false
+    alter_database_stmt: false
+    drop_database_stmt: false
+    create_table_stmt: true
+    drop_table_stmt: false
+    drop_sequence_stmt: false
+    rename_table_stmt: false
+    create_view_stmt: false
+    create_sequence_stmt: false
+    create_index_stmt: true
+    drop_index_stmt: true
+    lock_tables_stmt: false
+    unlock_tables_stmt: false
+    cleanup_table_lock_stmt: false
+    repair_table_stmt: false
+    truncate_table_stmt: false
+    recover_table_stmt: false
+    flash_back_table_stmt: false
+    alter_table_option: true
+    alter_table_add_columns: true
+    alter_table_add_constraint: true
+    alter_table_drop_column: false
+    alter_table_drop_primary_key: false
+    alter_table_drop_index: true
+    alter_table_drop_foreign_key: false
+    alter_table_modify_column: true
+    alter_table_change_column: true
+    alter_table_rename_column: false
+    alter_table_rename_table: false
+    alter_table_alter_column: true
+    alter_table_lock: false
+    alter_table_algorithm: false
+    alter_table_rename_index: true
+    alter_table_force: false
+    alter_table_add_partitions: false
+    alter_table_coalesce_partitions: false
+    alter_table_drop_partition: false
+    alter_table_truncate_partition: false
+    alter_table_partition: false
+    alter_table_enable_keys: false
+    alter_table_disable_keys: false
+    alter_table_remove_partitioning: false
+    alter_table_with_validation: false
+    alter_table_without_validation: false
+    alter_table_secondary_load: false
+    alter_table_secondary_unload: false
+    alter_table_rebuild_partition: false
+    alter_table_reorganize_partition: false
+    alter_table_check_partitions: false
+    alter_table_exchange_partition: false
+    alter_table_optimize_partition: false
+    alter_table_repair_partition: false
+    alter_table_import_partition_tablespace: false
+    alter_table_discard_partition_tablespace: false
+    alter_table_alter_check: false
+    alter_table_drop_check: false
+    alter_table_import_tablespace: false
+    alter_table_discard_tablespace: false
+    alter_table_index_invisible: false
+    alter_table_order_by_columns: false
+    alter_table_set_ti_flash_replica: false
+allowed_dml:
+    select_stmt: true
+    union_stmt: true
+    load_data_stmt: false
+    insert_stmt: true
+    delete_stmt: false
+    update_stmt: true
+    show_stmt: true
+    split_region_stmt: false
+boolean_field_linter: true
+charset_linter: true
+column_comment_linter: true
+column_name_linter: true
+created_at_default_value_linter: true
+created_at_exists_linter: true
+created_at_type_linter: true
+destruct_linter: true
+float_double_linter: true
+foreign_key_linter: true
+id_exists_linter: true
+id_is_primary_linter: true
+id_type_linter: true
+index_length_linter: true
+index_name_linter: true
+keywords_linter: true
+not_null_linter: true
+table_comment_linter: true
+table_name_linter: true
+updated_at_exists_linter: true
+updated_at_type_linter: true
+updated_at_default_value_linter: true
+updated_at_on_update_linter: true
+varchar_length_linter: true
