+++ tools/cli/command/root_cmd.go
@@ -48,7 +48,7 @@ var (
 )
 
 // Cmds which not require login
-var loginWhiteListCmds = strings.Join([]string{"init", "parse", "version", "help"}, ",")
+var loginWhiteListCmds = strings.Join([]string{"init", "parse", "version", "help", "miglint"}, ",")
 
 // rootCmd represents the base command when called without any subcommands
 var RootCmd = &cobra.Command{
